refactor(console): refactor the components structure
erfactor the components structure
