+++ packages/console/src/pages/JwtClaims/MonacoCodeEditor/index.tsx
@@ -14,6 +14,8 @@ import * as styles from './index.module.scss';
 import type { IStandaloneCodeEditor, Model } from './type.js';
 import useEditorHeight from './use-editor-height.js';
 
+export type { Model } from './type.js';
+
 type Props = {
   className?: string;
   actions?: React.ReactNode;
@@ -26,10 +28,13 @@ function MonacoCodeEditor({ className, actions, models }: Props) {
   const editorRef = useRef<Nullable<IStandaloneCodeEditor>>(null);
 
   const [activeModelName, setActiveModelName] = useState<string>();
+
   const activeModel = useMemo(
     () => models.find((model) => model.name === activeModelName),
     [activeModelName, models]
   );
+
+  const isMultiModals = useMemo(() => models.length > 1, [models]);
 
   const { containerRef, editorHeight } = useEditorHeight();
 
@@ -80,29 +85,31 @@ function MonacoCodeEditor({ className, actions, models }: Props) {
   return (
     <div className={classNames(className, styles.codeEditor)}>
       <header>
-        {models.length > 1 ? (
-          <div className={styles.tabList}>
-            {models.map(({ name, title, icon }) => (
-              <div
-                key={name}
-                className={classNames(styles.tab, name === activeModelName && styles.active)}
-                role="button"
-                tabIndex={0}
-                onClick={() => {
+        <div className={styles.tabList}>
+          {models.map(({ name, title, icon }) => (
+            <div
+              key={name}
+              className={classNames(
+                styles.tab,
+                isMultiModals && styles.tabButton,
+                name === activeModelName && styles.active
+              )}
+              {...(isMultiModals && {
+                role: 'button',
+                tabIndex: 0,
+                onClick: () => {
                   setActiveModelName(name);
-                }}
-                onKeyDown={onKeyDownHandler(() => {
+                },
+                onKeyDown: onKeyDownHandler(() => {
                   setActiveModelName(name);
-                })}
-              >
-                {icon}
-                {title}
-              </div>
-            ))}
-          </div>
-        ) : (
-          <div className={styles.title}>{activeModel?.title}</div>
-        )}
+                }),
+              })}
+            >
+              {icon}
+              {title}
+            </div>
+          ))}
+        </div>
         <div className={styles.actions}>
           {actions}
           <IconButton size="small" onClick={handleCodeCopy}>
