+++ packages/console/src/pages/JwtClaims/SettingsPanel/GuideCard.tsx
