+++ packages/console/src/pages/JwtClaims/SettingsPanel/index.module.scss
